adição do view e mudanças pontuais

+++ model/triangulo.py
@@ -8,7 +8,7 @@ class Triangulo(Figura):
     x2: int
     y2: int
     x3: float = (x2 + x1) / 2
-    y3: int = y2
+    y3: float = y2
 
     def incompleta(self):
         return (self.x1 == self.x2) or (self.y1 == self.y2)
